feat: fix return type from contractsEvents (#92)
* feat: fix return type from contractsEvents

* chore: comment to remove later
